render 404
when the link referenced by a given id is not found, we would want to
let the user know about this, the best way to do this would be to
return an error in json format to the user.

if the mix env is dev, however, we might not see the error response
since debug_errors in the dev.exs config file is turned on. Turning
it off converts Ecto.NoResultsError into a 404 response and renders
the 404.json to the user.